Struct libfirm_rs::nodes::Mulh

pub struct Mulh(_);

returns the upper word of the product of its operands (the part which would not fit into the result mode of a normal Mul anymore)

Methods

impl Mulh

pub fn left(self) -> Node

Gets first operand.

pub fn set_left(
self,
val: impl NodeTrait
)

Sets first operand.

pub fn right(self) -> Node

Gets second operand.

pub fn set_right(
self,
val: impl NodeTrait
)

Sets second operand.