Trait libfirm_rs::nodes::NodeTrait

pub trait NodeTrait {
    fn internal_ir_node(&self) -> *mut ir_node;

    fn as_node(&self) -> Node { ... }

    fn keep_alive(&self) { ... }

    fn mode(&self) -> Mode { ... }

    fn block(&self) -> Block { ... }

    fn set_block(&self, block: Block) { ... }

    fn out_nodes(&self) -> OutNodeIterator { ... }

    fn out_nodes_ex(&self) -> OutNodeExIterator { ... }

    fn all_out_projs(&self) -> Vec<Proj> { ... }

    fn collect_all_out_projs(&self, projs: &mut Vec<Proj>) { ... }

    fn in_nodes(&self) -> InNodeIterator { ... }

    fn set_input_at(&self, idx: i32, pred: Node) { ... }

    fn set_in_nodes(&self, nodes: &[Node]) { ... }

    fn node_id(&self) -> i64 { ... }

    fn is_pinned(&self) -> bool { ... }

    fn is_commutative(&self) -> bool { ... }

    fn is_only_valid_in_start_block(&self) -> bool { ... }

    fn graph(&self) -> Graph { ... }

    fn walk<F>(&self, walker: F)
    where
        F: FnMut(VisitTime, Node),
        Self: Sized,
    { ... }

    fn walk_dfs_in_block_stop_at_phi_node<Callback>(
        &self, 
        block: Block, 
        callback: &mut Callback
    )
    where
        Callback: FnMut(Node),
        Self: Sized,
    { ... }
}

A trait to abstract from Node enum and various *-Node structs. Inspired by https://github.com/libfirm/jFirm/blob/master/src/firm/nodes/Node.java.

Required methods

fn internal_ir_node(&self) -> *mut ir_node

Loading content...
Provided methods

fn as_node(&self) -> Node

fn keep_alive(&self)

fn mode(&self) -> Mode

fn block(&self) -> Block

fn set_block(&self, block: Block)

Important traits for OutNodeIterator
impl Iterator for OutNodeIterator type Item = Node;

fn out_nodes(&self) -> OutNodeIterator

Important traits for OutNodeExIterator
impl Iterator for OutNodeExIterator type Item = (Node, i32);

fn out_nodes_ex(&self) -> OutNodeExIterator

fn all_out_projs(&self) -> Vec<Proj>

fn collect_all_out_projs(&self, projs: &mut Vec<Proj>)

Important traits for InNodeIterator
impl Iterator for InNodeIterator type Item = Node;

fn in_nodes(&self) -> InNodeIterator

fn set_input_at(&self, idx: i32, pred: Node)

fn set_in_nodes(&self, nodes: &[Node])

fn node_id(&self) -> i64

fn is_pinned(&self) -> bool

fn is_commutative(&self) -> bool

fn is_only_valid_in_start_block(&self) -> bool

fn graph(&self) -> Graph

fn walk<F>(&self, walker: F) where
F: FnMut(VisitTime, Node),
Self: Sized,

libifrm irg_walk wrapper

Walks over the ir graph, starting at the this node and going to all predecessors, i.e., dependencies (operands) of this node. Note that this traversal crosses block boundaries, since blocks are also just predecessors in the Graph.

fn walk_dfs_in_block_stop_at_phi_node<Callback>(
&self,
block: Block,
callback: &mut Callback
) where
Callback: FnMut(Node),
Self: Sized,

Perform a DFS over all nodes within block starting at self. As soon as a Phi node is reached, that branch of the DFS is canceled. There is no callback for a Phi node. The primary use case for this API is in codegen.

Loading content...