Struct libfirm_rs::nodes_gen::Block

pub struct Block(_);

A basic block

Methods

impl Block

Important traits for CfgPredsIterator
impl Iterator for CfgPredsIterator type Item = Node;

pub fn cfg_preds(self) -> CfgPredsIterator

impl Block

pub fn entity(self) -> Entity

Gets entity representing this block.

pub fn set_entity(self, val: Entity)

Sets entity representing this block.

impl Block

pub fn new_add(self, irn_left: Node, irn_right: Node) -> Add

Creates a new Add-node.

irn_left left

irn_right right

pub fn new_alloc(
self,
irn_mem: Node,
irn_size: Node,
alignment: c_uint
) -> Alloc

Creates a new Alloc-node.

irn_mem mem

irn_size size

alignment alignment of the memory block (must be a power of 2)

pub fn new_and(self, irn_left: Node, irn_right: Node) -> And

Creates a new And-node.

irn_left left

irn_right right

pub fn new_bitcast(self, irn_op: Node, mode: *mut ir_mode) -> Bitcast

Creates a new Bitcast-node.

irn_op op

mode mode of the operations result

pub fn new_builtin(
self,
irn_mem: Node,
in_: Vec<Node>,
kind: Type,
ty: *mut ir_type
) -> Builtin

Creates a new Builtin-node.

irn_mem mem

in_ additional inputs

kind kind of builtin

ty method type for the builtin call

pub fn new_call(
self,
irn_mem: Node,
irn_ptr: Node,
in_: Vec<Node>,
ty: *mut ir_type
) -> Call

Creates a new Call-node.

irn_mem mem

irn_ptr ptr

in_ additional inputs

ty type of the call (usually type of the called procedure)

pub fn new_cmp(self, irn_left: Node, irn_right: Node, relation: Type) -> Cmp

Creates a new Cmp-node.

irn_left left

irn_right right

relation Comparison relation

pub fn new_cond(self, irn_selector: Node) -> Cond

Creates a new Cond-node.

irn_selector selector

pub fn new_confirm(
self,
irn_value: Node,
irn_bound: Node,
relation: Type
) -> Confirm

Creates a new Confirm-node.

irn_value value

irn_bound bound

relation relation of value to bound

pub fn new_conv(self, irn_op: Node, mode: *mut ir_mode) -> Conv

Creates a new Conv-node.

irn_op op

mode mode of the operations result

pub fn new_copyb(
self,
irn_mem: Node,
irn_dst: Node,
irn_src: Node,
ty: *mut ir_type,
flags: Type
) -> CopyB

Creates a new CopyB-node.

irn_mem mem

irn_dst dst

irn_src src

ty type of copied data

flags specifies volatility

pub fn new_div(
self,
irn_mem: Node,
irn_left: Node,
irn_right: Node,
pinned: i32
) -> Div

Creates a new Div-node.

irn_mem mem

irn_left left

irn_right right

pinned pinned state

pub fn new_eor(self, irn_left: Node, irn_right: Node) -> Eor

Creates a new Eor-node.

irn_left left

irn_right right

pub fn new_free(self, irn_mem: Node, irn_ptr: Node) -> Free

Creates a new Free-node.

irn_mem mem

irn_ptr ptr

pub fn new_ijmp(self, irn_target: Node) -> IJmp

Creates a new IJmp-node.

irn_target target

pub fn new_jmp(self) -> Jmp

Creates a new Jmp-node.

pub fn new_load(
self,
irn_mem: Node,
irn_ptr: Node,
mode: *mut ir_mode,
ty: *mut ir_type,
flags: Type
) -> Load

Creates a new Load-node.

irn_mem mem

irn_ptr ptr

mode mode of the value to be loaded

ty The type of the object which is stored at ptr (need not match with mode)

flags specifies alignment, volatility and pin state

pub fn new_member(self, irn_ptr: Node, entity: Entity) -> Member

Creates a new Member-node.

irn_ptr ptr

entity entity which is selected

pub fn new_minus(self, irn_op: Node) -> Minus

Creates a new Minus-node.

irn_op op

pub fn new_mod(
self,
irn_mem: Node,
irn_left: Node,
irn_right: Node,
pinned: i32
) -> Mod

Creates a new Mod-node.

irn_mem mem

irn_left left

irn_right right

pinned pinned state

pub fn new_mul(self, irn_left: Node, irn_right: Node) -> Mul

Creates a new Mul-node.

irn_left left

irn_right right

pub fn new_mulh(self, irn_left: Node, irn_right: Node) -> Mulh

Creates a new Mulh-node.

irn_left left

irn_right right

pub fn new_mux(self, irn_sel: Node, irn_false: Node, irn_true: Node) -> Mux

Creates a new Mux-node.

irn_sel sel

irn_false false

irn_true true

pub fn new_not(self, irn_op: Node) -> Not

Creates a new Not-node.

irn_op op

pub fn new_or(self, irn_left: Node, irn_right: Node) -> Or

Creates a new Or-node.

irn_left left

irn_right right

pub fn new_phi(self, in_: Vec<Node>, mode: *mut ir_mode) -> Phi

Creates a new Phi-node.

in_ additional inputs

mode mode of the operations result

pub fn new_pin(self, irn_op: Node) -> Pin

Creates a new Pin-node.

irn_op op

pub fn new_raise(self, irn_mem: Node, irn_exo_ptr: Node) -> Raise

Creates a new Raise-node.

irn_mem mem

irn_exo_ptr exo_ptr

pub fn new_return(self, irn_mem: Node, in_: Vec<Node>) -> Return

Creates a new Return-node.

irn_mem mem

in_ additional inputs

pub fn new_sel(self, irn_ptr: Node, irn_index: Node, ty: *mut ir_type) -> Sel

Creates a new Sel-node.

irn_ptr ptr

irn_index index

ty array type

pub fn new_shl(self, irn_left: Node, irn_right: Node) -> Shl

Creates a new Shl-node.

irn_left left

irn_right right

pub fn new_shr(self, irn_left: Node, irn_right: Node) -> Shr

Creates a new Shr-node.

irn_left left

irn_right right

pub fn new_shrs(self, irn_left: Node, irn_right: Node) -> Shrs

Creates a new Shrs-node.

irn_left left

irn_right right

pub fn new_store(
self,
irn_mem: Node,
irn_ptr: Node,
irn_value: Node,
ty: *mut ir_type,
flags: Type
) -> Store

Creates a new Store-node.

irn_mem mem

irn_ptr ptr

irn_value value

ty The type of the object which is stored at ptr (need not match with value's type)

flags specifies alignment, volatility and pin state

pub fn new_sub(self, irn_left: Node, irn_right: Node) -> Sub

Creates a new Sub-node.

irn_left left

irn_right right

pub fn new_switch(
self,
irn_selector: Node,
n_outs: c_uint,
table: *mut ir_switch_table
) -> Switch

Creates a new Switch-node.

irn_selector selector

n_outs number of outputs (including pn_Switch_default)

table table describing mapping from input values to Proj numbers

pub fn new_sync(self, in_: Vec<Node>) -> Sync

Creates a new Sync-node.

in_ additional inputs

pub fn new_tuple(self, in_: Vec<Node>) -> Tuple

Creates a new Tuple-node.

in_ additional inputs