Struct libfirm_rs::Graph

pub struct Graph { /* fields omitted */ }

Methods

impl Graph

pub fn function_with_entity(entity: Entity, num_slots: usize) -> Graph

pub fn entity(self) -> Entity

pub fn start_block(self) -> Block

pub fn end_block(self) -> Block

pub fn start(self) -> Start

pub fn set_start_block(self, block: Block)

pub fn end(self) -> End

pub fn args(self) -> Proj

pub fn no_mem(self) -> NoMem

pub fn frame(self) -> Node

pub fn dump(self, suffix: &str)

pub fn assure_outs(self)

pub fn remove_bads(self)

pub fn walk_topological<F>(self, walker: F) where
F: FnMut(&Node),

Walks over all reachable nodes in the graph, ensuring that nodes inside a basic block are visited in topological order.

Nodes in different blocks might get visited in an interleaved order.

Parameters

walker walker function

Does not use the link field.

pub fn nodes(self) -> Vec<Node>

pub fn exchange(
prev: impl NodeTrait,
new: impl NodeTrait
)

pub fn exchange_value(
prev: impl ValueNode + Into<Node>,
new: impl ValueNode + Into<Node>
)

pub fn mark_as_bad(
self,
node: impl NodeTrait
)

Replace the given node with a "bad" node, thus marking it and all the nodes dominated by it as unreachable. The whole subtree can then be removed using Graph::remove_bads.

pub fn copy_node<F>(self, node: Node, copy_fn: F) -> Node where
F: FnMut(Node) -> Node,

pub fn new_imm_block(self, preds: &[Node]) -> Block

pub fn slots(self) -> i32

pub fn finalize_construction(self)

impl Graph

pub fn new_add(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Add

Creates a new Add-node.

block The block.

irn_left left

irn_right right

pub fn new_address(self, entity: Entity) -> Address

Creates a new Address-node.

entity entity to operate on

pub fn new_align(self, mode: Mode, ty: Ty) -> Align

Creates a new Align-node.

mode mode of the operations result

ty type to operate on

pub fn new_alloc(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_size: impl NodeTrait,
alignment: c_uint
) -> Alloc

Creates a new Alloc-node.

block The block.

irn_mem mem

irn_size size

alignment alignment of the memory block (must be a power of 2)

pub fn new_and(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> And

Creates a new And-node.

block The block.

irn_left left

irn_right right

pub fn new_bad(self, mode: Mode) -> Bad

Creates a new Bad-node.

mode mode of the operations result

pub fn new_bitcast(
self,
block: Block,
irn_op: impl NodeTrait,
mode: Mode
) -> Bitcast

Creates a new Bitcast-node.

block The block.

irn_op op

mode mode of the operations result

pub fn new_block(self, in_: &[Node]) -> Block

Creates a new Block-node.

in_ additional inputs

pub fn new_builtin(
self,
block: Block,
irn_mem: impl NodeTrait,
in_: &[Node],
kind: Type,
ty: Ty
) -> Builtin

Creates a new Builtin-node.

block The block.

irn_mem mem

in_ additional inputs

kind kind of builtin

ty method type for the builtin call

pub fn new_call(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait,
in_: &[Node],
ty: Ty
) -> Call

Creates a new Call-node.

block The block.

irn_mem mem

irn_ptr ptr

in_ additional inputs

ty type of the call (usually type of the called procedure)

pub fn new_cmp(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait,
relation: Type
) -> Cmp

Creates a new Cmp-node.

block The block.

irn_left left

irn_right right

relation Comparison relation

pub fn new_cond(
self,
block: Block,
irn_selector: impl NodeTrait
) -> Cond

Creates a new Cond-node.

block The block.

irn_selector selector

pub fn new_confirm(
self,
block: Block,
irn_value: impl NodeTrait,
irn_bound: impl NodeTrait,
relation: Type
) -> Confirm

Creates a new Confirm-node.

block The block.

irn_value value

irn_bound bound

relation relation of value to bound

pub fn new_const(self, tarval: Tarval) -> Const

Creates a new Const-node.

tarval constant value (a tarval object)

pub fn new_conv(
self,
block: Block,
irn_op: impl NodeTrait,
mode: Mode
) -> Conv

Creates a new Conv-node.

block The block.

irn_op op

mode mode of the operations result

pub fn new_copyb(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_dst: impl NodeTrait,
irn_src: impl NodeTrait,
ty: Ty,
flags: Type
) -> CopyB

Creates a new CopyB-node.

block The block.

irn_mem mem

irn_dst dst

irn_src src

ty type of copied data

flags specifies volatility

pub fn new_div(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait,
pinned: i32
) -> Div

Creates a new Div-node.

block The block.

irn_mem mem

irn_left left

irn_right right

pinned pinned state

pub fn new_dummy(self, mode: Mode) -> Dummy

Creates a new Dummy-node.

mode mode of the operations result

pub fn new_end(self, in_: &[Node]) -> End

Creates a new End-node.

in_ additional inputs

pub fn new_eor(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Eor

Creates a new Eor-node.

block The block.

irn_left left

irn_right right

pub fn new_free(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait
) -> Free

Creates a new Free-node.

block The block.

irn_mem mem

irn_ptr ptr

pub fn new_ijmp(
self,
block: Block,
irn_target: impl NodeTrait
) -> IJmp

Creates a new IJmp-node.

block The block.

irn_target target

pub fn new_jmp(self, block: Block) -> Jmp

Creates a new Jmp-node.

block The block.

pub fn new_load(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait,
mode: Mode,
ty: Ty,
flags: Type
) -> Load

Creates a new Load-node.

block The block.

irn_mem mem

irn_ptr ptr

mode mode of the value to be loaded

ty The type of the object which is stored at ptr (need not match with mode)

flags specifies alignment, volatility and pin state

pub fn new_member(
self,
block: Block,
irn_ptr: impl NodeTrait,
entity: Entity
) -> Member

Creates a new Member-node.

block The block.

irn_ptr ptr

entity entity which is selected

pub fn new_minus(
self,
block: Block,
irn_op: impl NodeTrait
) -> Minus

Creates a new Minus-node.

block The block.

irn_op op

pub fn new_mod(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait,
pinned: i32
) -> Mod

Creates a new Mod-node.

block The block.

irn_mem mem

irn_left left

irn_right right

pinned pinned state

pub fn new_mul(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Mul

Creates a new Mul-node.

block The block.

irn_left left

irn_right right

pub fn new_mulh(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Mulh

Creates a new Mulh-node.

block The block.

irn_left left

irn_right right

pub fn new_mux(
self,
block: Block,
irn_sel: impl NodeTrait,
irn_false: impl NodeTrait,
irn_true: impl NodeTrait
) -> Mux

Creates a new Mux-node.

block The block.

irn_sel sel

irn_false false

irn_true true

pub fn new_nomem(self) -> NoMem

Creates a new NoMem-node.

pub fn new_not(
self,
block: Block,
irn_op: impl NodeTrait
) -> Not

Creates a new Not-node.

block The block.

irn_op op

pub fn new_offset(self, mode: Mode, entity: Entity) -> Offset

Creates a new Offset-node.

mode mode of the operations result

entity entity to operate on

pub fn new_or(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Or

Creates a new Or-node.

block The block.

irn_left left

irn_right right

pub fn new_phi(self, block: Block, in_: &[Node], mode: Mode) -> Phi

Creates a new Phi-node.

block The block.

in_ additional inputs

mode mode of the operations result

pub fn new_pin(
self,
block: Block,
irn_op: impl NodeTrait
) -> Pin

Creates a new Pin-node.

block The block.

irn_op op

pub fn new_proj(
self,
irn_pred: impl NodeTrait,
mode: Mode,
num: c_uint
) -> Proj

Creates a new Proj-node.

irn_pred pred

mode mode of the operations result

num number of tuple component to be extracted

pub fn new_raise(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_exo_ptr: impl NodeTrait
) -> Raise

Creates a new Raise-node.

block The block.

irn_mem mem

irn_exo_ptr exo_ptr

pub fn new_return(
self,
block: Block,
irn_mem: impl NodeTrait,
in_: &[Node]
) -> Return

Creates a new Return-node.

block The block.

irn_mem mem

in_ additional inputs

pub fn new_sel(
self,
block: Block,
irn_ptr: impl NodeTrait,
irn_index: impl NodeTrait,
ty: Ty
) -> Sel

Creates a new Sel-node.

block The block.

irn_ptr ptr

irn_index index

ty array type

pub fn new_shl(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Shl

Creates a new Shl-node.

block The block.

irn_left left

irn_right right

pub fn new_shr(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Shr

Creates a new Shr-node.

block The block.

irn_left left

irn_right right

pub fn new_shrs(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Shrs

Creates a new Shrs-node.

block The block.

irn_left left

irn_right right

pub fn new_size(self, mode: Mode, ty: Ty) -> Size

Creates a new Size-node.

mode mode of the operations result

ty type to operate on

pub fn new_start(self) -> Start

Creates a new Start-node.

pub fn new_store(
self,
block: Block,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait,
irn_value: impl NodeTrait,
ty: Ty,
flags: Type
) -> Store

Creates a new Store-node.

block The block.

irn_mem mem

irn_ptr ptr

irn_value value

ty The type of the object which is stored at ptr (need not match with value's type)

flags specifies alignment, volatility and pin state

pub fn new_sub(
self,
block: Block,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Sub

Creates a new Sub-node.

block The block.

irn_left left

irn_right right

pub fn new_switch(
self,
block: Block,
irn_selector: impl NodeTrait,
n_outs: c_uint,
table: *mut ir_switch_table
) -> Switch

Creates a new Switch-node.

block The block.

irn_selector selector

n_outs number of outputs (including pn_Switch_default)

table table describing mapping from input values to Proj numbers

pub fn new_sync(self, block: Block, in_: &[Node]) -> Sync

Creates a new Sync-node.

block The block.

in_ additional inputs

pub fn new_tuple(self, block: Block, in_: &[Node]) -> Tuple

Creates a new Tuple-node.

block The block.

in_ additional inputs

pub fn new_unknown(self, mode: Mode) -> Unknown

Creates a new Unknown-node.

mode mode of the operations result