Struct libfirm_rs::nodes::Block

pub struct Block(_);

A basic block

Methods

impl Block

pub fn deepest_common_dominator(a: Self, b: Self) -> Self

pub fn immediate_dominator(self) -> Option<Self>

pub fn loop_depth(self) -> u32

pub fn immediate_post_dominator(self) -> Option<Self>

Important traits for CfgPredsIterator
impl Iterator for CfgPredsIterator type Item = Node;

pub fn cfg_preds(self) -> CfgPredsIterator

Important traits for PhisOfBlockLinkedListIterator
impl Iterator for PhisOfBlockLinkedListIterator type Item = Phi;

pub fn linked_list_of_phis(self) -> PhisOfBlockLinkedListIterator

Access the phis of a block using get_Block_phis and get_Phi_next. Note that this list is not updated automatically. Fill the list manually or use phis instead.

pub fn phis(self) -> Vec<Phi>

pub fn phi_or_node(self, nodes: &[Node]) -> Node

pub fn set_cfg_pred(
self,
idx: i32,
pred: impl NodeTrait
)

pub fn mature(self)

pub fn value(self, slot_idx: usize, mode: Mode) -> Node

pub fn set_value(
self,
slot_idx: usize,
val: impl NodeTrait
)

pub fn cur_store(self) -> Node

pub fn set_store(
self,
s: impl NodeTrait
)

pub fn imm_add_pred(
self,
pred: impl NodeTrait
)

pub fn all_nodes_in_block(
self
) -> impl Iterator<Item = Node>

pub fn dom_depth(self) -> usize

pub fn dominates(self, other: Self) -> bool

impl Block

pub fn entity(self) -> Entity

Gets entity representing this block.

pub fn set_entity(self, val: Entity)

Sets entity representing this block.

impl Block

pub fn new_add(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Add

Creates a new Add-node.

irn_left left

irn_right right

pub fn new_alloc(
self,
irn_mem: impl NodeTrait,
irn_size: impl NodeTrait,
alignment: c_uint
) -> Alloc

Creates a new Alloc-node.

irn_mem mem

irn_size size

alignment alignment of the memory block (must be a power of 2)

pub fn new_and(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> And

Creates a new And-node.

irn_left left

irn_right right

pub fn new_bitcast(
self,
irn_op: impl NodeTrait,
mode: Mode
) -> Bitcast

Creates a new Bitcast-node.

irn_op op

mode mode of the operations result

pub fn new_builtin(
self,
irn_mem: impl NodeTrait,
in_: &[Node],
kind: Type,
ty: Ty
) -> Builtin

Creates a new Builtin-node.

irn_mem mem

in_ additional inputs

kind kind of builtin

ty method type for the builtin call

pub fn new_call(
self,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait,
in_: &[Node],
ty: Ty
) -> Call

Creates a new Call-node.

irn_mem mem

irn_ptr ptr

in_ additional inputs

ty type of the call (usually type of the called procedure)

pub fn new_cmp(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait,
relation: Type
) -> Cmp

Creates a new Cmp-node.

irn_left left

irn_right right

relation Comparison relation

pub fn new_cond(
self,
irn_selector: impl NodeTrait
) -> Cond

Creates a new Cond-node.

irn_selector selector

pub fn new_confirm(
self,
irn_value: impl NodeTrait,
irn_bound: impl NodeTrait,
relation: Type
) -> Confirm

Creates a new Confirm-node.

irn_value value

irn_bound bound

relation relation of value to bound

pub fn new_conv(
self,
irn_op: impl NodeTrait,
mode: Mode
) -> Conv

Creates a new Conv-node.

irn_op op

mode mode of the operations result

pub fn new_copyb(
self,
irn_mem: impl NodeTrait,
irn_dst: impl NodeTrait,
irn_src: impl NodeTrait,
ty: Ty,
flags: Type
) -> CopyB

Creates a new CopyB-node.

irn_mem mem

irn_dst dst

irn_src src

ty type of copied data

flags specifies volatility

pub fn new_div(
self,
irn_mem: impl NodeTrait,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait,
pinned: i32
) -> Div

Creates a new Div-node.

irn_mem mem

irn_left left

irn_right right

pinned pinned state

pub fn new_eor(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Eor

Creates a new Eor-node.

irn_left left

irn_right right

pub fn new_free(
self,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait
) -> Free

Creates a new Free-node.

irn_mem mem

irn_ptr ptr

pub fn new_ijmp(
self,
irn_target: impl NodeTrait
) -> IJmp

Creates a new IJmp-node.

irn_target target

pub fn new_jmp(self) -> Jmp

Creates a new Jmp-node.

pub fn new_load(
self,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait,
mode: Mode,
ty: Ty,
flags: Type
) -> Load

Creates a new Load-node.

irn_mem mem

irn_ptr ptr

mode mode of the value to be loaded

ty The type of the object which is stored at ptr (need not match with mode)

flags specifies alignment, volatility and pin state

pub fn new_member(
self,
irn_ptr: impl NodeTrait,
entity: Entity
) -> Member

Creates a new Member-node.

irn_ptr ptr

entity entity which is selected

pub fn new_minus(
self,
irn_op: impl NodeTrait
) -> Minus

Creates a new Minus-node.

irn_op op

pub fn new_mod(
self,
irn_mem: impl NodeTrait,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait,
pinned: i32
) -> Mod

Creates a new Mod-node.

irn_mem mem

irn_left left

irn_right right

pinned pinned state

pub fn new_mul(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Mul

Creates a new Mul-node.

irn_left left

irn_right right

pub fn new_mulh(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Mulh

Creates a new Mulh-node.

irn_left left

irn_right right

pub fn new_mux(
self,
irn_sel: impl NodeTrait,
irn_false: impl NodeTrait,
irn_true: impl NodeTrait
) -> Mux

Creates a new Mux-node.

irn_sel sel

irn_false false

irn_true true

pub fn new_not(
self,
irn_op: impl NodeTrait
) -> Not

Creates a new Not-node.

irn_op op

pub fn new_or(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Or

Creates a new Or-node.

irn_left left

irn_right right

pub fn new_phi(self, in_: &[Node], mode: Mode) -> Phi

Creates a new Phi-node.

in_ additional inputs

mode mode of the operations result

pub fn new_pin(
self,
irn_op: impl NodeTrait
) -> Pin

Creates a new Pin-node.

irn_op op

pub fn new_raise(
self,
irn_mem: impl NodeTrait,
irn_exo_ptr: impl NodeTrait
) -> Raise

Creates a new Raise-node.

irn_mem mem

irn_exo_ptr exo_ptr

pub fn new_return(
self,
irn_mem: impl NodeTrait,
in_: &[Node]
) -> Return

Creates a new Return-node.

irn_mem mem

in_ additional inputs

pub fn new_sel(
self,
irn_ptr: impl NodeTrait,
irn_index: impl NodeTrait,
ty: Ty
) -> Sel

Creates a new Sel-node.

irn_ptr ptr

irn_index index

ty array type

pub fn new_shl(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Shl

Creates a new Shl-node.

irn_left left

irn_right right

pub fn new_shr(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Shr

Creates a new Shr-node.

irn_left left

irn_right right

pub fn new_shrs(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Shrs

Creates a new Shrs-node.

irn_left left

irn_right right

pub fn new_store(
self,
irn_mem: impl NodeTrait,
irn_ptr: impl NodeTrait,
irn_value: impl NodeTrait,
ty: Ty,
flags: Type
) -> Store

Creates a new Store-node.

irn_mem mem

irn_ptr ptr

irn_value value

ty The type of the object which is stored at ptr (need not match with value's type)

flags specifies alignment, volatility and pin state

pub fn new_sub(
self,
irn_left: impl NodeTrait,
irn_right: impl NodeTrait
) -> Sub

Creates a new Sub-node.

irn_left left

irn_right right

pub fn new_switch(
self,
irn_selector: impl NodeTrait,
n_outs: c_uint,
table: *mut ir_switch_table
) -> Switch

Creates a new Switch-node.

irn_selector selector

n_outs number of outputs (including pn_Switch_default)

table table describing mapping from input values to Proj numbers

pub fn new_sync(self, in_: &[Node]) -> Sync

Creates a new Sync-node.

in_ additional inputs

pub fn new_tuple(self, in_: &[Node]) -> Tuple

Creates a new Tuple-node.

in_ additional inputs